Add helper for voting scokets logic
